@@ -1,6 +1,6 @@
 {
   "name": "@stellar/js-xdr",
-  "version": "3.1.1",
+  "version": "3.1.2",
   "description": "Read/write XDR encoded data structures (RFC 4506)",
   "main": "lib/xdr.js",
   "browser": "dist/xdr.js",
@@ -48,10 +48,10 @@
     "ui": "bdd"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.1",
-    "@babel/eslint-parser": "^7.24.1",
-    "@babel/preset-env": "^7.24.1",
-    "@babel/register": "^7.23.7",
+    "@babel/core": "^7.24.9",
+    "@babel/eslint-parser": "^7.24.8",
+    "@babel/preset-env": "^7.24.8",
+    "@babel/register": "^7.24.6",
     "babel-loader": "^9.1.3",
     "buffer": "^6.0.3",
     "chai": "^4.3.10",
@@ -70,12 +70,12 @@
     "karma-sinon-chai": "^2.0.2",
     "karma-webpack": "^5.0.1",
     "lint-staged": "13.2.2",
-    "mocha": "^10.3.0",
+    "mocha": "^10.6.0",
     "prettier": "^2.8.7",
     "sinon": "^15.2.0",
     "sinon-chai": "^3.7.0",
     "terser-webpack-plugin": "^5.3.10",
-    "webpack": "^5.90.3",
+    "webpack": "^5.93.0",
     "webpack-cli": "^5.0.2"
   }
 }
