@@ -1,6 +1,6 @@
 {
   "name": "@stellar/stellar-base",
-  "version": "12.0.1",
+  "version": "12.1.0",
   "description": "Low-level support library for the Stellar network.",
   "main": "./lib/index.js",
   "browser": {
@@ -70,15 +70,15 @@
   },
   "homepage": "https://github.com/stellar/js-stellar-base",
   "devDependencies": {
-    "@babel/cli": "^7.24.7",
-    "@babel/core": "^7.24.7",
-    "@babel/eslint-parser": "^7.24.7",
+    "@babel/cli": "^7.24.8",
+    "@babel/core": "^7.24.9",
+    "@babel/eslint-parser": "^7.24.8",
     "@babel/eslint-plugin": "^7.24.7",
-    "@babel/preset-env": "^7.24.7",
+    "@babel/preset-env": "^7.24.8",
     "@babel/register": "^7.24.6",
     "@definitelytyped/dtslint": "^0.0.182",
     "@istanbuljs/nyc-config-babel": "3.0.0",
-    "@types/node": "^20.14.2",
+    "@types/node": "^20.14.11",
     "@typescript-eslint/parser": "^6.20.0",
     "babel-loader": "^9.1.3",
     "babel-plugin-istanbul": "^6.1.1",
@@ -91,7 +91,7 @@
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prefer-import": "^0.0.1",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^5.2.1",
     "eslint-webpack-plugin": "^4.2.0",
     "ghooks": "^2.0.4",
     "husky": "^8.0.3",
@@ -105,22 +105,22 @@
     "karma-webpack": "^5.0.1",
     "lint-staged": "^15.2.7",
     "minami": "^1.1.1",
-    "mocha": "^10.4.0",
+    "mocha": "^10.6.0",
     "node-polyfill-webpack-plugin": "^3.0.0",
     "nyc": "^15.1.0",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "randombytes": "^2.1.0",
     "sinon": "^16.1.0",
     "sinon-chai": "^3.7.0",
     "taffydb": "^2.7.3",
     "terser-webpack-plugin": "^5.3.10",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "webpack": "^5.92.0",
+    "typescript": "^5.5.3",
+    "webpack": "^5.93.0",
     "webpack-cli": "^5.1.1"
   },
   "dependencies": {
-    "@stellar/js-xdr": "^3.1.1",
+    "@stellar/js-xdr": "^3.1.2",
     "base32.js": "^0.1.0",
     "bignumber.js": "^9.1.2",
     "buffer": "^6.0.3",
